Refactor to separate standby frequency component
